server/requirements.txt
aiohttp==3.9.5
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
fastapi==0.115.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.10
google-api-core==2.19.2
google-api-python-client==2.185.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.63.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.63.0rc1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-google-genai==2.0.5
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.147
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.9.0
numpy==1.26.4
orjson==3.11.4
packaging==24.2
pandas==2.3.3
pillow==10.4.0
pinecone==5.0.0
pinecone-client==5.0.0
pinecone-plugin-inference==1.0.2
pinecone-plugin-interface==0.0.7
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.2
pydantic-settings==2.11.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.28.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.66.5
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.30.6
watchdog==5.0.3
yarl==1.22.0
